jobs/serializers.py
# jobs/serializers.py

from rest_framework import serializers
from .models import Job

class JobSerializer(serializers.ModelSerializer):
    class Meta:
        model = Job
        fields = ['id', 'title', 'company', 'location', 'description', 'date_posted', 'posted_by']
        read_only_fields = ('posted_by',)